PAVITHRA_V/Loan Approval System/requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-groq
python-dotenv
typesense
langchain_openai
langgraph
crewai
crewai_tools
load_dotenv
langchain-huggingface